requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
google-play-scraper==1.2.7
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
